Add files via upload
Evolution of international flights originating in the US using monthly and annual summaries from the last 27 yeasrs.
Both graphs shows a constant increase of air traffic leaving the US. The included data accounts not only human movement but also freight and mail exchange. From the monthly a drop in traffic after a sudden pick at the end and beginning of each year during the last 20 years can be identified.
The graph generated from the annual summary shows two mayor drops in air traffic from 2000 to 2010. The first one seems to be related to 9/11 which led to a 15% decrease the following year in air traffic but after one year the annual increase came back to its historical average until the economical crisis around 2007 - 2008 that produced a 12.5% decrease in traffic. However is interesting to notice that after each crisis, the aeronautical industry is capable to recover its growth trends.
It would be interesting to analyze the behavior of passenger, freight and mail separately to see the travelers reaction to destabilizing world events.