Add files via upload
Top 15 destinations from US originating traffic.
The graphs include data from passengers, freight and mail. It doesn't come to a surprise that the mayor US exchanges include Canada and Mexico and the UK occupying the top 3 places during the last 27 years. However it seems that it took 6 to 8 years after the NAFTA agreement to see almost an 80% increase between the US and its two most important regional partners, Mexico and Canada.
It is also interesting to notice that after the war in Irak, opposed by France,  the US exchanges with this country slightly decreased and from being in top 5 countries  of favorite destinations, it passed to position number ten in the new century.
Finally, we can also see the top 15 destinations the continuous presence of mayor US allies in different parts of the word such as Colombia, Germany, Japan and South Korea. However for this countries it would be more interesting to study if its constant appearance in the top 15 it is due to visitors or to freight. For example, Colombia having a violent image but being an important hub for South America, I will tend to say that its appearance in the top 15 corresponds to mail and freight exchange. Analyzing the behavior of passengers, freight and mail separately will give more insights on this.